+++ packages/shared/components/popups/Index.svelte
@@ -8,7 +8,6 @@
     import AddNode from './AddNode.svelte'
     import AirdropNetworkInfo from './AirdropNetworkInfo.svelte'
     import Backup from './Backup.svelte'
-    import BalanceFinder from './BalanceFinder.svelte'
     import Busy from './Busy.svelte'
     import CrashReporting from './CrashReporting.svelte'
     import CreateAccountPopup from './CreateAccountPopup.svelte'
@@ -40,6 +39,7 @@
     import SwitchNetwork from './SwitchNetwork.svelte'
     import Version from './Version.svelte'
     import Video from './Video.svelte'
+    import WalletFinderPopup from './WalletFinderPopup.svelte'
     import ConfirmDeveloperProfile from './ConfirmDeveloperProfile.svelte'
     import LegalUpdate from './LegalUpdate.svelte'
     import SingleAccountGuide from './SingleAccountGuide.svelte'
@@ -123,7 +123,7 @@
         diagnostics: Diagnostics,
         riskFunds: RiskFunds,
         missingBundle: MissingBundle,
-        walletFinder: BalanceFinder,
+        walletFinder: WalletFinderPopup,
         snapshot: Snapshot,
         video: Video,
         // Participation (voting / staking)
