+++ packages/shared/routes/dashboard/settings/views/advanced/WalletFinder.svelte
