+++ packages/shared/routes/dashboard/settings/views/advanced/index.js
@@ -1,4 +1,3 @@
-export { default as BalanceFinder } from './BalanceFinder.svelte'
 export { default as CrashReporting } from './CrashReporting.svelte'
 export { default as DeepLinks } from './DeepLinks.svelte'
 export { default as Diagnostics } from './Diagnostics.svelte'
@@ -6,3 +5,4 @@ export { default as ErrorLog } from './ErrorLog.svelte'
 export { default as HiddenAccounts } from './HiddenAccounts.svelte'
 export { default as MigrateLedgerIndex } from './MigrateLedgerIndex.svelte'
 export { default as NetworkConfiguration } from './NetworkConfiguration.svelte'
+export { default as WalletFinder } from './WalletFinder.svelte'
